requirements.txt
# Appy Pie Automate Toolkit — Dependencies
# https://appypieautomate.ai

streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
